+++ acfml.php
@@ -11,6 +11,8 @@
 
 if( !defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
+use ACFML\ACF_Multilingual;
+
 define( 'ACFML', true );
 define( 'ACFML_PATH', plugin_dir_path( __FILE__ ) );
 define( 'ACFML_BASENAME', plugin_basename( __FILE__ ) );
@@ -18,7 +20,7 @@ define( 'ACFML_URL', plugins_url('/', __FILE__) );
 
 
 require_once(ACFML_PATH . 'vendor/autoload.php');
-require_once(ACFML_PATH . 'inc/class.acfml.php');
+require_once(ACFML_PATH . 'inc/class.acf-multilingual.php');
 
 /**
  * acfml