+++ inc/class.admin.php
@@ -14,7 +14,7 @@ class Admin {
    * @param ACF_Multilingual|null $acfml
    * @author Rasso Hilber <mail@rassohilber.com>
    */
-  public function __construct(?\ACF_Multilingual $acfml = null) {
+  public function __construct(ACF_Multilingual $acfml) {
 
     // inject main class
     $this->acfml = $acfml;