+++ inc/class.post-types-controller.php
@@ -30,7 +30,7 @@ class Post_Types_Controller {
    * @param ACF_Multilingual|null $acfml
    * @author Rasso Hilber <mail@rassohilber.com>
    */
-  public function __construct(?\ACF_Multilingual $acfml = null) {
+  public function __construct(ACF_Multilingual $acfml) {
     
     // inject main class
     $this->acfml = $acfml;