+++ inc/class.taxonomies-controller.php
@@ -21,7 +21,7 @@ class Taxonomies_Controller {
    * @param ACF_Multilingual|null $acfml
    * @author Rasso Hilber <mail@rassohilber.com>
    */
-  public function __construct(?\ACF_Multilingual $acfml = null) {
+  public function __construct(ACF_Multilingual $acfml) {
 
     // inject main class
     $this->acfml = $acfml;